func/requirements.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
heimdall-client==2.2.0
etria-logger==0.1.8
mnemosine==0.1.2
aioboto3==9.6.0
flask==2.0.3
pydantic==1.9.0
python-decouple==3.6
hypercorn==0.13.2
